Removed ServerBenchmark
Given that prune now is testing performance quite well, and appears to
be able to reliably pick up and identify regressions in many places, and
also given that we we haven't run the benchmark in a long time, I've
removed the ServerBenchmark to simplify the build.

Also removed the NettyHandlerSpec, which depended on the same
NettyRunner that the ServerBenchmark depended on - all the tests in
there were covered by the other tests that made real HTTP connections to
the server, and so didn't just test that we were calling what we thought
we were supposed to be calling on Netty, but tested that we were calling
the right methods on Netty.

The reason I've deleted these is because they were getting quite
expensive to maintain - there was a custom implementation of a router,
custom implementation of play.api.Application, etc, and I was finding in
the implementation of many of the DI changes in Play, there was always
something in these benchmarks that was breaking that needed to be fixed,
because these tests weren't reusing any of the existing test
infrastructure provided by Play.