Swap first AkkaHttpServerSpec test and increase timeout
The first test in AkkaHttpServerSpec has been failing in master. The
purpose of this change is to investigate why. Is it only the first
test that fails? Is the failure something to do with the slow CI test
box? Once we've worked out the solution, we should be able to revert
this change.