Upgrade to jboss-logging 3.2.0
This is to avoid the LGPL license for jboss-logging 3.1.x which is
brought in by hibernate-validator.  I did a manual audit of the changes
from 3.1.1 to 3.2.0, the only binary incompatible change was the
addition of methods to an interface, but it's not likely that anything
we use or that our users use would be implementing this interface.