Don't add the Set-Cookie header unless there are cookies
The Set-Cookie header was being set for every Java result, whether or not it
had any cookies. If there weren't any cookies it was set to an invalid empty
value. This problem wasn't visible in Netty because there is code in
NettyResultStreamer to split Set-Cookie headers into multiple headers. This
code automatically removed empty headers. This fix is needed to get Akka HTTP
to work properly.