Ensure filename headers are correctly skipped
Fixes #3814

There were actually two bugs, the first was that the null character
after the filename (and comment) wasn't dropped, the second is that the
dropWhile iteratee implementation was completely messed up if the
filename was split over multiple chunks.

I fixed both by replacing the dropWhile iteratee with a
dropWhileIncluding iteratee - it could have been done with a dropWhile +
drop, but the dropWhileIncluding can be a little more efficient.