Use trampoline to avoid Cont.flatMap stack overflows
Fixes #3855.

The closure inside Iteratee.flatMap could call itself recursively,
leading to a StackOverflowError when Cont is flatMapped thousands of
times. An unresolved problem is the fact that this closure can still
grow so large that it could exhaust memory. This is hard to avoid
because there is no easy way in Scala to optimize recursive function
composition.