Before exitValue() on forked process, wait a bit
If waiting a bit fails, then also process.destroy() because
the forked process is not exiting and must be punished.

Also, move joining the output-logger threads to after
the process exits, so we are more likely to catch any
interesting output, and to be sure those threads aren't
blocking on the output pipes from the process.