fork-run/dev-mode: assert we're talking to the right app
Our controller now returns "original-fork-run" and we explode
if we get another value, such as from another scripted test's app,
which is what's happening at the moment.

We can't check the reload count until after we verify we're talking
to the right app, or else we get a confusing error that reloads are
broken when in fact the problem is the wrong app.

Also, lengthen the timeouts to 1250 instead of 1000 - if OS resolution is 1000,
we should be a little longer.