Permit importing Play Java project with build errors in Eclipse. Fix #3818
When generating the Eclipse files for a Play Java project, the SCALA_CONTAINER
classpath entry (generated by sbteclipse) needs to be replaced with the plain
Scala library. This is done because Java developers may not have the Scala IDE
plugin for Eclipse installed (the SCALA_CONTAINER is a Scala IDE container, not
available with the standard Eclipse installation).

To carry out the above, the location of the scala-library.jar needs to be located.
This used to be achieved by running `dependencyClasspath in Runtime`. The issue with
this command is that it will force compilation of the project, and hence it will
fail to execute if the project doesn't compile. Changing this command to
`dependencyClasspath in Compile` no longer forces compilation of the project.

With the change proposed in this commit, generating the Eclipse files for either
a Scala or Java Play project will work exactly the same.