Return 500 response if header's field is invalid. Fixes #3950
When mapping the Play `Result`'s headers to the netty `DefaultHttpResponse`
(i.e., `nettyResponse`), an `RuntimeException` can occur if any of the
headers' field name contains an invalid [1] character.

Prior to this commit, such an exception would be handled by closing the
current connection and sending no answer back to the client, giving the
impression that the server hang (as reported in the ticket). See
https://github.com/playframework/playframework/blob/master/framework/src/play-netty-server/src/main/scala/play/core/server/netty/NettyResultStreamer.scala#L92

The implemented solution is to fail gracefully when an header field name
contains an invalid character, and return a 500 result (Internal Server
Error).

An alternative could have been to validate the headers immediately when set in
the Play `ResponseHeader` (and making the condition a class' invariant). In
fact, that may be a more elegant way of fixing the issue reported in #3950.

[1] See section "4.2 Message Headers” in http://www.ietf.org/rfc/rfc2616.txt