Set Content-Length header for fixed-length Results
When we have Writeable content we know the length of the result
body, so we can easily set the Content-Length header. The header
would usually be set anyway after the ServerResultUtils reads
the Enumerator and determines that the content has a know length.
Setting the header early saves the effort of reading the enumerator
and allows the result to be sent earlier.