Load server configuration from standard reference.conf files
Previously it was loaded from server-reference.conf files, but
I think we can simplify our code and use the same loading code
for both the server and the application. The difference is that
the configuration in dev mode will have access to the
application.conf file, whereas the server will not. In prod
mode they will both have access to application.conf.