Fixed fork run protocol compile flakiness
A race condition in the Scala compiler that relates to global state and
compiling the same code at the same time with two different instances of
the Scala compiler causes intermittent build failures.  The existing
work around that we had in place was to take the compile and doc tasks
of the fork run protocol projects so they would not be done concurrently
with each other.  SBT 0.13.8 split the compile task into multiple
smaller tasks, meaning that the task that we now have to tag is
compileIncremental.