Allow lang to be set transiently
Currently in the Java API it's impossible to set the current
Context's language without also setting a cookie. That means
that it's impossible to just change the language for a single
request. This change adds new necessary methods to support
changing and clearing the language without changing the
cookie.

Other changes:
* Added unit tests for changing the language.
* Added documentation tests for Java internationalization.