Expose PlayJsonModule as public API
This allows people to use Jackson in interesting ways with JsValue
serialisers/deserialisers.

I also moved all the jackson related code to a package of its own
(everything was private to play before so this is not a breaking change)
and I removed a stored static reference to the context classloader which
was possibly a source of a classloader leak.  The classloader was not
used anywhere, and was possibly left over from an earlier requirement.