Documentation for migrating `Plugin` to `Module`
Code snippets for the migration documentation has been placed in proper scala
and java files because we want these sources to be compiled in the 2.4.x branch.
When backporting this commit, we will also have to add a further commit to edit
the Build.scala as follow:

```
PlayDocsKeys.javaManualSourceDirectories :=
  (baseDirectory.value / "manual" / "working" / "javaGuide" ** codeFilter).get ++
  (baseDirectory.value / "manual" / "releases" / "migration24" ** codeFilter).get,
PlayDocsKeys.scalaManualSourceDirectories :=
  (baseDirectory.value / "manual" / "working" / "scalaGuide" ** codeFilter).get ++
  (baseDirectory.value / "manual" / "releases" / "migration24" ** codeFilter).get,
```

Which basically ensures both Scala and Java sources located under the
migration24 folder are compiled.