Replaced `F.Predicate` with Java8 `java.util.function.Predicate`
As explained in d24c79d85eb9f71499588d2180bfb29e6b73d55f, we are embracing Java8
functional interfaces. This commit replaces `F.Predicate` with
`java.util.function.Predicate`.

The change is source compatible if the lambda's body passed to `F.Predicate`
does not throw a checked exception. Otherwise, users will need to update their
code to catch the exception and provide the necessary logic to deal with it.