Replaced `F.Function` with Java8 friends. Fix #4690
As explained in d24c79d85eb9f71499588d2180bfb29e6b73d55f, we are embracing Java8
functional interfaces. This commit makes the following changes:

* `F.Function0` -> `java.util.function.Supplier`
* `F.Function1` -> `java.util.function.Function`
* `F.Function2` -> `java.util.function.BiFunction`

The change is source compatible if the lambda's body passed to either of the
former `F.Function` types does not throw a checked exception. Otherwise, users
will need to update their code to catch the exception and provide the necessary
logic to deal with it.

There is however an important change affecting source compatibility, as several
of the methods in `F.Promise` used to throw (Throwable), while they don't now (
Ref #4691).

There is one additional function type that has no corresponding in the Java8
library: `F.Function3`. This type is currently only used in the `RoutingDsl`, so
I initially thought we could remove it. However, it's tricky to remove
functionality, so I decided to leave it alone for the moment, as we can always
revisit this in the future.