Make JSON macros return OFormat/OWrites.
The JSON macros remove the boilerplate for creating Reads/Writes/Format for case classes.

But given that a case class will always serialise to a JsObject then .writes/.format could return
OWrites/OFormat respectively.

This allows for type-safety in manipulating a statically known JsObject, rather than a JsValue.

Given that OWrites and OFormat are subtypes of Writes/Format this is a backwards-compatible change.