Bumped akka-stream and akka-http to 1.0-RC4. Fix #4774
* Moving from RC3 to RC4 had quite an impact because of the many class'
  renamings.

* It seems there is an overloading issue with `Flow.to` (see comment in
  `play.libs.streams.Accumulator.through`). We should report the problem
  to the akka team.

* Had to make `play.it.http.JavaResultsHandlingSpec` and
  `play.it.http.ScalaResultsHandlingSpec` running sequentially because they
  would fail otherwise (not sure what has changed in akka-http to cause it, but
  I assumed it may have been caused by additional asynchronism in the impl).

One thing that worries me is that some akka-streams type leak into the Play user
API (for instance, see `FlowMaterializer` in `play.api.mvc.Filters`), and
akka-streams is an experimental module that hence doesn't ensure backward binary
compatibility across releases. I'm not sure if this is actually going to be an
issue, but I think we should keep it in mind.