Use akka streams for WebSockets API.
**This is working in progress, so please don't merge it just yet.**

In this commit the signature of several WS API methods was changed and it's no longer source compatible. In the linked ticket, @jroper hinted that we should have provided overloads ffor the new methods. While that would be desiderable, it's unfortunately not achievable, because the erased signature of the overloads would be the same. I guess the only alternative is to rename the existing methods, deprecate them in favour of the new ones using akka stream's types, and document the changes in the migration guide. Since this is a change that may be controversial, I'd like to gather feedback before investing time in this task.

\cc @jroper @richdougherty @wsargent