Akka streams support for WebSockets
Fixes #4686.

I've simplified the WebSocket interface somewhat, so it's simply a
function that takes a request and returns a Flow[Message, Message, _].

FrameFormatter has been replaced with MessageFlowTransformer, which
works on both ends of the flow at the same time.  This is necessary
to allow the WebSocket to be closed with an appropriate error if the
wrong frame is sent.

This now allows the user to trigger and handle ping/pong messages, as
well as send custom close messages.

I also implemented support for WebSockets in the Akka HTTP backend.

What's not done:

* Still uses iteratees on top of Netty backend - this will be fixed when
  we upgrade to Netty 4 with Netty reactive streams.
* Documentation is not updated.