Preserve the order of loading enabled modules
Some third-party play modules are sensitive about the order of loading.
For example, a db migration module has to be loaded before a db fixture module is loaded.

Currently, the order of loading modules are destroyed by `toSet` to remove duplicate modules.
This commit changed to use `distinct` instead.
`distinct` removes duplicate modules and the result sequence consists of the first occurence.
This works as if it preserves the order of modules in most cases.