Update deployment documentation
Change main class assembly plugin documentation:

When using the assembly plugin to build an uberjar, the documentation says
`play.core.server.NettyServer` should be used as the main class. Doing so
results in an error message

https://github.com/playframework/playframework/blob/6606b051b0511898440739d2cd13a4cf5f7604fa/framework/src/play-netty-server/src/main/scala/play/core/server/NettyServer.scala#L216

Now the documentation contains the correct main class.

Add documentation for preventing PID file creation:

It is not always nice, that Play applications create their own PID. In many
cases this is even a problem. It is easy to prevent Play from doing this and
extending the documentation on this, will hopefully help people to get Play
running the way they want.