Added async streaming to the Java WS API. Fix #4973
Introduced a `stream()` method in the Java WS API that closely emulates its
Scala counterpart. The only difference is that instead of returning a `Tuple2`,
it returns a `StreamedResponse`. I felt this is more idiomatic for Java users,
and I'm actually of the idea that the same should be done in the Scala API. So,
I went ahead and did it.

Currently, the Java WS streaming functionality relies on the Scala one, which is
based on iteratees. The plan is to change this and use akka stream, but #4952
needs to be addressed first.

I've also restored the old signature of the WS Scala API methods that were
deprecated in commit 6edaedfe69184214a06521b681442f50973dc704