Bumped HikariCP version to 2.4.1
HikariCP 2.4.x stream will be binary compatible with 2.4.1 as mentioned in
https://groups.google.com/forum/#!topic/play-framework-dev/iqKgOqa3McM.

The changelog for this release can be read here:
https://groups.google.com/forum/#!topic/hikari-cp/FGygaj_yrxg

Also, note that HikariCP 2.4.0 inadvertently broke binary compatibility and
hence was flagged as a bad release.

Finally, the `HikariDataSource.shutdown` method was deprecated, so I've replaced
it with the `close` method.