Made it possible to write body parers in Java
Fixes #4733 #1732

Refactored the Java BodyParser interface so it can be implemented by
Java users to actually parse a body, rather than just wrap a Scala body
parser. The parsers are injected and instantiated by Guice - this
required a small amount of refactoring around anything that handles Java
actions.

The Java RequestBody class was refactored such that the body is passed
into it as an Object, and inferred based on type when as* methods are
called.

I didn't implement support for custom handling of multipart/form-data
file parts, since that is dependent on #5037 and tracked in #5044.