Forcing shutdown of AHC event loop to prevent OOME: cannot create new threads
After upgrading AHC to version 2.0.0-alpha11 (because it supports reactive
stream, which we need in the next commit), the tests suite started to fail with
a "OutOfMemoryException: cannot create new native thread". This happened because
this new version of AHC is shutting down the event loop gracefully, i.e., it
allows some time between the moment a client closes the AHC client, and the
moment when the related AHC event loop is actually shut down (this is to allow
that in-flight messages have a chance to be received).

While this behavior makes sense in the real world, it's causing problems in the
Play tests suite. This is because several tests using the AHC client are run in
parallel, leading to OutOfMemoryException because threads pool from tests that
have already terminated are left around because AHC is shutting them down
gracefully.

In this commit we are restoring the old AHC behavior, i.e., the event loop is
shut down as soon as the AHC `close` method is called. Though, I feel the right
fix is to make the parameters that handle the AHC event loop's shutdown
configurable, so that different defaults can be used for testing.

See also AsyncHttpClient/async-http-client#969