Updated Netty version used by Play's WS test client to Netty 3.10.4
Play's websocket test client still uses Netty 3.x, which has a different
artifact name to Netty 4.x. The upgrade to async-http-client 2 removed our last
dependency on Netty 3.x - except, selenium was bringing in a very old version of
Netty, 3.5.2, which is why everything still compiled. I'm guessing that the
problem is that the very old Netty version has a bug in it that was triggering
it to not shutdown properly. Using 3.10.4 seems to fix the problem.