Bumped async-http-client version to 2.0.0-alpha12
The Proxi API has changed a bit in this async-http-client (AHC) release.
Specifically, `HTTPS` protocol can no longer be specified for a proxy instance.
The reason is that  now you can configure both regular and secured ports on a
single instance, hence the `HTTPS` protocol is not needed anymore.

An additional feature offered by AHC is that it now allows to configure different
ports for HTTP and HTTPS. Exposing this feature in Play would require us to
change the Play `WSProxyServer` and add a new member to configure a secure port.
This should be straightforward to implement, but I'm not sure we need it, so I
haven't done it.