Removed used of `sequential` in many Spec test classes
I believe we are overusing the specs2 `sequential` keyword, which forces
examples (i.e., tests) within a Spec class to be ran sequentially. Ideally, we
should allow for as much parallelism as possible, as this should reduce the time
required to run the test suite, and it may help us catch race conditions and
other concurrency hiccups that are otherwise hard to reproduce.

Furthermore, forcing a sequential execution where it's not needed also sends the
wrong message to contributors, and incourage writing tests that cannot be run in
parallel (see the changes in this commit to `ResultSpec`).

Finally, note that the changes in this commit do not affect the way tests are
run in the CI server, because we are already forcing them to run on a single CPU
(check the .travis file, we are passing `"set concurrentRestrictions in Global
+= Tags.limitAll(1)"` to sbt).