Generalized stop hook callback return type
In most places the stop hook callback return type was already generalized to
`Future[_]`, so this commit is only finishing the work and updating the few
places where `Future[Unit]` was still used.

The change is source compatible. Furthermore, it removes the need of having to
map futures when their inner type isn't `Unit`.