Update Java seed template
This updates the Java seed template with the same set of changes that
I recently made to the Scala seed template (#5528). As part of the
change I moved some common code into the `templates/play-common`
directory.

Both the Java and Scala templates still need their tests to be
updated to current best practices, e.g. using mocks, dependency
injection, use ScalaTest, etc. This is something I plan to do in a
subsequent PR.

A couple of other misc changes:
- fix Java Filter implementation
- bind ExecutionContextExecutor in BuiltInModule