Fix broken docs build in downstream projects
The addition of commonManualSourceDirectories with no default value
broke all the downstream projects that use the PlayDocsPlugin (breaking
the nightly build, for example).

This provides a default value for it, additionally it also provides
default values for the java and scala source directories, and the
migration sources.  It also makes the migrationManualSources setting
public so it can actually be set, and removes non side effecting setting
of it that did nothing.

In fixing this, I inadvertantly fixed a build issue in the documentation
where a lot of sources weren't being compiled/tested. Of course, those
sources were horribly broken, so I had to fix them.