Partially restore BodyParsers.parse.maxLength behaviour WRT 2.4.x
Ensure that when handling a `MaxLengthLimitAttained` error that the
relevant limit is the same as that expected. Otherwise, the limit
given to the `maxLength` body parser is handled erroneously by that
of the underlying parser, resulting in a 413 response being returned.

Note that if the `maxLength` parser and the underlying parser have
the same limit, the underlying one will 'win', returning 413.