Rewrote maxLength body parser
Fixes #5704

Rather than signalling max length exceeded by terminating the stream
with a failure, this new parser materialises the status of the max
length check so that it can be worked with explicitly.  This means that
if multiple max length flows are applied to the same stream, they don't
impact one another.