Removed iteratees and many deprecated methods
* Iteratees now live at https://github.com/playframework/play-iteratees
* Iteratees reactive streams implemntation was also moved there
* Removed all deprecated methods/classes that use iteratees
* Removed all deprecated Java methods of handling streams (Chunks,
  LegacyWebSocket)
* Removed F.Promise
* Removed GlobalSettings
* Removed most deprecated methods from play.Play/play.api.Play (only
  left methods for accessing the application).

One thing that this required was updating the WebSockets documentation -
that part of this PR would be useful to backport to 2.5.

Since a number of APIs were removed, documentation links to non existent
scala/javadocs, particularly in old migration guides, had to be removed,
because they would now become dead.