Added note on LoggerConfigurator in AppLoader.
The current version of SettingsLogger for 2.5.x creates the impression that in
any case, adding a logback config file is all that is required in order to
override Play's default logging behaviour. The exception, however, is when
one provides a dedicated application logger. This is indirectly mentioned
in the Migration25 documentation, but as my own hour-long search shows, it's
probably worth to make this clear here as well, as this provides a more
thorough coverage of all things logging.