Move environment definition into code example (#6148)
Since environment is referenced in the parser definition, it needs to be included in the documented snippet of code. It took me a few minutes of head scratching to work out what this should be, and I was heartened to see that I ended up with the exact same definition as is in the code, it's just not included in the code sample on the page.