Better Result error handling (#6153)
Fixes #6099.

When the backend server (Netty, Akka HTTP) discovers an error in a
result (e.g. invalid headers) the server needs to abort processing the
result and send an error result to the user. Previously, this error
handling was done with custom error handling code for the server.

This pull request changes the server behavior to use the application's
HttpErrorHandler to generate the error result instead of generating a
custom error response. Using the existing error handler makes the
error handling more standardized. It also means that the application
will be notified when an error occurs during result generation, which
will make it easier to track errors that occur in Play applications
(see #6098, #6099).