Add tests of ServerResultUtils#validateResult to make sure an empty body is sent for 1xx, 204 and 304 responses
Purpose
Currently there's no tests to check if the returned body of 1xx, 204 and 304 responses
is empty, which is required by RFC7230. To make sure that no entity is sent to a client,
we add unit tests of ServerResultUtils#validateResult[1] that takes responsibility for
cancelling a body if necessary.

Background Context
PR(#6261)[2] was merged to meet the requirement of RFC7230. BasicHttpClient is used for
the integration tests to check if an empty body is returned for 1xx, 204 and 304 responses.
These integration tests, however, don't guarantee that those responses return an empty body
as our BasicHttpClient just returns an empty body instead of sending an actual body.[3]

References
PR(#6261)[2]
PR Comment[4]

[1] https://github.com/playframework/playframework/blob/2.5.4/framework/src/play-server/src/main/scala/play/core/server/common/ServerResultUtils.scala#L67-L68
[2] https://github.com/playframework/playframework/pull/6261
[3] https://github.com/playframework/playframework/blob/2.5.4/framework/src/play-integration-test/src/test/scala/play/it/http/BasicHttpClient.scala#L214-L218
[4] https://github.com/playframework/playframework/pull/6261#issuecomment-235596925