Support typed attributes for requests (#6321)
RequestHeader and Request can now have custom attributes added to
them. This lets plugins and users add their own data to requests.
Adding custom attributes was sort of possible by using tags, but tag
values could only be strings. The new API adds support for typed
attributes.

I have intentionally restricted the API to hide internal details of
how attributes are implemented. This is because I want to make it
possible to change the underlying implementation later if we want to
support more complicated features. In particular, I've avoided exposing
the underlying map that's used to store the attributes. This is
because we might want to support lazy or dynamic attributes later, and
these types of attributes might not use a map.

Fixes #5814.

Other misc changes:
* Changed Java Request's username property to be an use attribute.
* Refactored Java's Request to implement it in Scala. Fixes #6193.
* Use `Handler.Stage` for request and handler rewriting.