Update ScalaForms documentation with options example
Add an example for using the options helper to create a select.
Previously, it wasn't clear to people unfamiliar with the frameworks
helpers that a pair passed to the options helper would result in having
the first element of said pair be the value of the generated mark up and
the second being the displayed value (in the case of an option).