Move code examples to external files
Moving code examples out to external files so that they can be compiled
and ensured in their validity. Changed code to be something small to a
more complete example of using the options and the select helpers.

Adding test to show that values of a FormField are used to select the
selected value in a select helper based on their values matching the
first value in a pair given to the options helper.