Update ScalaForms documentation with options example (#6654)
* Update ScalaForms documentation with options example

Add an example for using the options helper to create a select.
Previously, it wasn't clear to people unfamiliar with the frameworks
helpers that a pair passed to the options helper would result in having
the first element of said pair be the value of the generated mark up and
the second being the displayed value (in the case of an option).

* Move code examples to external files

Moving code examples out to external files so that they can be compiled
and ensured in their validity. Changed code to be something small to a
more complete example of using the options and the select helpers.

Adding test to show that values of a FormField are used to select the
selected value in a select helper based on their values matching the
first value in a pair given to the options helper.